docs: Update Readme.md and commitMessage.ts
In Readme.md, the description of the available commands was updated. The commands `init` and `uninit` were removed, and the description for `generate-commit` was updated to reflect that it is currently the only command available for managing conventional commits.

In commitMessage.ts, a line of code that was previously used to create a folder structure if the file was not found in .diff was removed. Now, if the file is not found, a message is logged and a 'no diff found' file is created, but no additional folders are created.